requirements.txt
pysnowflake==0.1.3
fastapi==0.109.0
pandas==2.2.0
pymongo==4.6.1